move status update to test_utils + move out of test_only as many fns as possible
